+++ src/lib/Scenes/City/City.tsx
@@ -9,7 +9,6 @@ import { EventEmitter } from "../Map/EventEmitter"
 import { Tab } from "../Map/types"
 import { AllEvents } from "./Components/AllEvents"
 import { CityTab } from "./Components/CityTab"
-
 interface Props {
   verticalMargin?: number
   isDrawerOpen?: boolean
@@ -78,34 +77,36 @@ export class CityView extends Component<Props, State> {
     return (
       buckets && (
         <Theme>
-          <Box>
-            <Flex py={1} alignItems="center">
-              <Handle />
-            </Flex>
-            <FiltersBar
-              tabs={this.filters}
-              goToPage={activeIndex => EventEmitter.dispatch("filters:change", activeIndex)}
-            />
-            <ScrollView
-              contentInset={{ bottom: bottomInset }}
-              onLayout={layout => (this.scrollViewVerticalStart = layout.nativeEvent.layout.y)}
-              scrollEnabled={isDrawerOpen}
-              ref={r => {
-                if (r) {
-                  this.scrollView = r as any
-                }
-              }}
-            >
-              {(() => {
-                switch (filter && filter.id) {
-                  case "all":
-                    return <AllEvents cityName={cityName} currentBucket={filter.id as BucketKey} buckets={buckets} />
-                  default:
-                    return <CityTab bucket={buckets[filter.id]} type={filter.text} relay={this.state.relay} />
-                }
-              })()}
-            </ScrollView>
-          </Box>
+          <>
+            <Box>
+              <Flex py={1} alignItems="center">
+                <Handle />
+              </Flex>
+              <FiltersBar
+                tabs={this.filters}
+                goToPage={activeIndex => EventEmitter.dispatch("filters:change", activeIndex)}
+              />
+              <ScrollView
+                contentInset={{ bottom: bottomInset }}
+                onLayout={layout => (this.scrollViewVerticalStart = layout.nativeEvent.layout.y)}
+                scrollEnabled={isDrawerOpen}
+                ref={r => {
+                  if (r) {
+                    this.scrollView = r as any
+                  }
+                }}
+              >
+                {(() => {
+                  switch (filter && filter.id) {
+                    case "all":
+                      return <AllEvents cityName={cityName} currentBucket={filter.id as BucketKey} buckets={buckets} />
+                    default:
+                      return <CityTab bucket={buckets[filter.id]} type={filter.text} relay={this.state.relay} />
+                  }
+                })()}
+              </ScrollView>
+            </Box>
+          </>
         </Theme>
       )
     )