+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - Adds CityTab stying and passes CityName to AllEvents view - kieran
 - Fixes some analytics tracking events - luc
 - Displays ShowCard when pin is tapped on Map view - luc
+- Adds styled city picker modal - ashley
 
 ### 1.8.7